Add capability to compute.limits object to poll other projects
At present, it is not possible for an admin user to query the limits of
all projects using this SDK.

This patch enables a client with valid admin credentials to pass the
project_id parameter to compute.limits.get to query projects other than
their own.

Example use of this: an admin generating a list of all the quotas and
usages of all projects in the cloud.

Change-Id: I0043697da8abde9950e04f14802e9f752459e7b9